requirements.txt
streamlit==1.22.0
SpeechRecognition==3.10.0
groq==0.4.9
python-dotenv==1.0.0
nltk==3.8.1
scikit-learn==1.2.2
numpy==1.24.3
pandas==2.0.1
matplotlib==3.7.1
seaborn==0.12.2
pyannote.audio==2.1.1
textblob==0.17.1
aiohttp==3.8.4
beautifulsoup4==4.12.2
cachetools==5.3.0
aiolimiter==1.1.0
plotly==5.14.1
spacy==3.5.3